Update upgrade-buildpacks pipeline
This change splits upgrade-* into stage-* and promote-*.

Stage will create a new buildpack with the name _latest if it doesn't
exist, or update the existing one if _latest does exist.

Promote will promote the _latest buildpack to be the current active
one, leaving no _latest buildpack.

Promote is not set to trigger by default, allowing the operator to
test a new buildpack before promoting it. One caveat here is that
while one buildpack is under test a new one may be found and uploaded,
replacing the new staged buildpack. The way around this is to pause
the stage job for that buildpack until testing is complete, then
promote, then unpause the stage job.

[#139704745]