Create release pushes to different resource
I thought I could be saved from having to do this with [ci skip], but it
turns out that is not the case. The dependent get when doing a put
causes Concourse to ignore the [ci skip] rule and see a new version
anyway.

See related issue: https://github.com/concourse/git-resource/issues/18

This reverts commit 08837a67b9965cb31c8a017de19e7ddc14da8a81.