Change how install-pcf handles terraform state
Previously create-infrastructure would create a terraform-state file and
only put it to the s3 resource if it was successful; in the failure case
it would try to roll the changes back.

There was a bug in the rollback after a refactor was done that caused
rollback to fail.

This has all been changed so that it will always put to the s3 resource,
whether creating infrastructure was successful or not, and it will
additionally use the tfstate from any prior run as the basis for
creating infrastructure the second time around. This means you can
safely retry create-infrastructure after fixing any params issues that
caused the failure.

It does this by adding a `get: terraform-state` to the job; this
presents some difficulty because initially there is no state. Thus a new
job, create-initial-terraform-state, was created. This will create an
empty file in the s3 resource that can be pulled down by
create-infrastructure. It is a no-op if the file already exists.

Because we'd like to be able to re-run create-infrastructure with any
updated terraform-state, we can't set a passed constraint saying it has
to have come after create-initial-terraform-state. This is unfortunate.